LayoutTests/crypto/subtle/hmac-check-algorithm-expected.txt

Test that HMAC operations only work when hash functions match between invocation and key.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a raw HMAC SHA-1 key from string literal...
PASS crypto.subtle.sign({name: 'hmac', hash: {name: 'sha-256'}}, key, asciiToUint8Array('foo')) threw exception Error: NotSupportedError: DOM Exception 9.
PASS crypto.subtle.verify({name: 'hmac', hash: {name: 'sha-256'}}, key, asciiToUint8Array('fake signature'), asciiToUint8Array('foo')) threw exception Error: NotSupportedError: DOM Exception 9.
PASS successfullyParsed is true

TEST COMPLETE
